Move the action cluster from Agent to a HasActions mixin
