+++ mesa/experimental/actions/__init__.py
@@ -1,8 +1,9 @@
-"""Scenarios module."""
+"""Timed, interruptible actions and the agent mixin that performs them."""
 
-from .actions import Action, ActionState
+from .actions import Action, ActionState, HasActions
 
 __all__ = [
     "Action",
     "ActionState",
+    "HasActions",
 ]
